+++ agent_generic_base/custom_model/custom.py
@@ -13,6 +13,11 @@
 # limitations under the License.
 # isort: off
 from helpers_telemetry import *  # noqa # pylint: disable=unused-import
+import os
+
+# Some libraries collect telemetry data by default. Let's disable that.
+os.environ["RAGAS_DO_NOT_TRACK"] = "true"
+os.environ["DEEPEVAL_TELEMETRY_OPT_OUT"] = "YES"
 # isort: on
 
 from agent import MyAgent