+++ agent_langgraph/agent_cli/kernel.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 import json
 import os
+import time
 from typing import Any, cast
 
+import requests
 from openai import OpenAI
 from openai.types.chat import (
     ChatCompletion,
@@ -123,12 +125,18 @@ class Kernel:
     @staticmethod
     def get_output(output_path: str) -> Any:
         """Read the local output file and remove it."""
-        with open(output_path, "r") as f:
-            output = f.read()
-
         if os.path.exists(output_path):
-            os.remove(output_path)
-        return output
+            with open(output_path, "r") as f:
+                output = f.read()
+
+            if os.path.exists(output_path):
+                os.remove(output_path)
+            return output
+        else:
+            print(
+                f"ERROR: Output file not found: {output_path}. Please check the agent execution logs for errors."
+            )
+            return None
 
     def local(
         self,
@@ -152,6 +160,61 @@ class Kernel:
             print(f"Error executing command: {e}")
             raise
 
+    def custom_model(self, custom_model_id: str, user_prompt: str) -> str:
+        chat_api_url = f"{self.base_url}/api/v2/genai/agents/fromCustomModel/{custom_model_id}/chat/"
+        print(chat_api_url)
+
+        headers = {
+            "Authorization": f"Bearer {os.environ['DATAROBOT_API_TOKEN']}",
+            "Content-Type": "application/json",
+        }
+        data = {"messages": [{"role": "user", "content": user_prompt}]}
+
+        print(f'Querying custom model with prompt: "{data}"')
+        print(
+            "Please wait... This may take 1-2 minutes the first time you run this as a codespace is provisioned "
+            "for the custom model to execute."
+        )
+        response = requests.post(
+            chat_api_url,
+            headers=headers,
+            json=data,
+        )
+
+        if not response.ok or not response.headers.get("Location"):
+            raise Exception(response.text)
+        # Wait for the agent to complete
+        status_location = response.headers["Location"]
+        while response.ok:
+            time.sleep(1)
+            response = requests.get(
+                status_location, headers=headers, allow_redirects=False
+            )
+            if response.status_code == 303:
+                agent_response = requests.get(
+                    response.headers["Location"], headers=headers
+                ).json()
+                # Show the agent response
+                break
+            else:
+                status_response = response.json()
+                if status_response["status"] in ["ERROR", "ABORTED"]:
+                    raise Exception(status_response)
+        else:
+            raise Exception(response.content)
+
+        if "errorMessage" in agent_response and agent_response["errorMessage"]:
+            return (
+                f"Error: "
+                f"{agent_response.get('errorMessage', 'No error message available')}"
+                f"Error details:"
+                f"{agent_response.get('errorDetails', 'No details available')}"
+            )
+        elif "choices" in agent_response:
+            return str(agent_response["choices"][0]["message"]["content"])
+        else:
+            return str(agent_response)
+
     def deployment(
         self, deployment_id: str, user_prompt: str, completion_json: str = ""
     ) -> ChatCompletion:
@@ -170,5 +233,9 @@ class Kernel:
         )
 
         print(f'Querying deployment with prompt: "{completion_create_params}"')
+        print(
+            "Please wait for the agent to complete the response. This may take a few seconds to minutes "
+            "depending on the complexity of the agent workflow."
+        )
         completion = openai_client.chat.completions.create(**completion_create_params)
         return completion